+++ Libplanet/Action/ActionEvaluationsExtensions.cs
@@ -22,8 +22,7 @@ namespace Libplanet.Action
                 .SelectMany(a => a.OutputStates.StateUpdatedAddresses)
                 .ToImmutableHashSet();
             IImmutableSet<(Address, Currency)> updatedFungibleAssets = actionEvaluations
-                .SelectMany(a => a.OutputStates.UpdatedFungibleAssets
-                    .SelectMany(kv => kv.Value.Select(c => (kv.Key, c))))
+                .SelectMany(a => a.OutputStates.UpdatedFungibleAssets)
                 .ToImmutableHashSet();
             IImmutableSet<Currency> updatedTotalSupplies = actionEvaluations
                 .SelectMany(a => a.OutputStates.TotalSupplyUpdatedCurrencies)